rum is great
some milestones here.
preferences can use some links, but they are fleshed out well.

rating system is not alive yet but easily remedied.

broadcasts and subscribes need some more thought but i think the reverse index and the straightaway sequence index are the right approach for easy comms & signals on event shifts

what elts...

need offline versions of pages for logged-in users because the web socket sets the ring req session var...

which means that locally persisted logins may be able to view everything because submission will work with a valid auth-token...

so just make all the threads visible who cares?

not a bad idea...

but the homepage demands a login.
i like it.
weird style but probably good